+++ CHANGELOG.rst
@@ -6,6 +6,10 @@ Forthcoming
 * Add intensity data to LaserScan
 * Print user message if dropped MonitoringFrames are detected
 * Print user message in case of timeout while waiting for MonitoringFrames
+* Adjust min and max scanner range. Fixes #70
+* Print diagnostic messages from PSENscan
+* Notify user about start retry
+* Add Continuous Acceptance Test
 * Contributors: Pilz GmbH and Co. KG
 
 0.1.0 (2020-10-19)