+++ CHANGELOG.rst
@@ -1,8 +1,8 @@
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 Changelog for package psen_scan_v2
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-Forthcoming
------------
+0.1.1 (2020-11-10)
+------------------
 * Add intensity data to LaserScan
 * Print user message if dropped MonitoringFrames are detected
 * Print user message in case of timeout while waiting for MonitoringFrames